+++ driver-core/CHANGELOG.rst
@@ -34,6 +34,8 @@ CHANGELOG
 - [bug] Fix issue when executing a PreparedStatement from another cluster
   (JAVA-641)
 - [improvement] Log keyspace xxx does not exist at WARN level (JAVA-534)
+- [improvement] Allow Cluster subclasses to delegate to another instance
+  (JAVA-619)
 
 Merged from 2.0.9_fixes branch: